debt(guards): give both CLI invocation guards one matcher (#1271)
The `gaia-ci-*` template reference guard decided whether a declared CLI
command still appeared in its template with a bare `includes('gaia ' +
command)` substring test. That requires a space immediately after the
binary name, so a template invoking a declared command through a quoted,
release-resolved path satisfies its contract in fact and fails it in the
check, reporting a reference as dropped from a template that still
invokes it.

The sibling reachability guard had already been widened for the same
assumption, and the invocation form that broke it there is a frozen
contract in this repository. Extract that matcher into
`util/gaia-invocation-matcher.ts` and point both guards at it, rather
than copying a second regex that would drift from the first silently.

The template guard's docblock recorded its one accepted gap in terms
broad enough to read as covering this check too, which is what left the
repair sitting as a fork. Scope it to the check it belongs to: the gap is
about extracting invocations nobody declared, and this check verifies a
string the contract already names.
